[Feat] 친구 검색 스크린 분리 및 검색 초기화 기능 구현 - 친구 검색 스크린으로 분리하여 친구 검색 탭 누르면 넘어가게 navigation, viewmodel도 넘어감 - 검색 창 지울 때 이전 검색 결과 초기화 되는 것 반영
