+++ app/src/main/java/com/and04/naturealbum/ui/NatureAlbumApp.kt
@@ -1,5 +1,6 @@
 package com.and04.naturealbum.ui
 
+import android.util.Log
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.runtime.Composable
@@ -7,8 +8,10 @@ import androidx.compose.runtime.remember
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
 import com.and04.naturealbum.ui.album.AlbumScreen
 import com.and04.naturealbum.ui.albumfolder.AlbumFolderScreen
+import com.and04.naturealbum.ui.friend.FriendSearchScreen
 import com.and04.naturealbum.ui.home.HomeScreen
 import com.and04.naturealbum.ui.labelsearch.LabelSearchScreen
 import com.and04.naturealbum.ui.maps.MapScreen
@@ -51,7 +54,7 @@ fun NatureAlbumApp(
 
         composable(NavigateDestination.SavePhoto.route) { backStackEntry ->
             val viewmodel = remember(backStackEntry) {
-                state.getNavBackStackEntry()
+                state.getNavBackStackEntry(NavigateDestination.SavePhoto.route)
             }
             SavePhotoScreen(
                 location = state.lastLocation.value,
@@ -68,7 +71,7 @@ fun NatureAlbumApp(
 
         composable(NavigateDestination.SearchLabel.route) { backStackEntry ->
             val viewmodel = remember(backStackEntry) {
-                state.getNavBackStackEntry()
+                state.getNavBackStackEntry(NavigateDestination.SavePhoto.route)
             }
 
             LabelSearchScreen(
@@ -107,11 +110,23 @@ fun NatureAlbumApp(
         }
 
         composable(NavigateDestination.MyPage.route) {
-            MyPageScreen(navigateToHome = { state.popupBackStack() })
+            MyPageScreen(
+                navigateToHome = { state.popupBackStack() },
+                navigateToFriendSearchScreen = { state.navigateToFriendSearch() },
+            )
         }
 
         composable(NavigateDestination.Map.route) {
             MapScreen(state.lastLocation.value)
+        }
+        composable(NavigateDestination.FriendSearch.route) { backStackEntry ->
+            val viewmodel = remember(backStackEntry) {
+                state.getNavBackStackEntry(NavigateDestination.MyPage.route)
+            }
+            FriendSearchScreen(
+                onBack = { state.popupBackStack() },
+                friendViewModel = hiltViewModel(viewmodel),
+            )
         }
     }
 }
