+++ app/src/main/java/com/and04/naturealbum/ui/mypage/MyPageItems.kt
@@ -4,7 +4,6 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
@@ -12,37 +11,20 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Search
 import androidx.compose.material3.Button
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.SearchBar
-import androidx.compose.material3.SearchBarDefaults
-import androidx.compose.material3.SuggestionChip
-import androidx.compose.material3.SuggestionChipDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import coil3.compose.AsyncImage
 import com.and04.naturealbum.R
 import com.and04.naturealbum.data.dto.FirebaseFriend
 import com.and04.naturealbum.data.dto.FirebaseFriendRequest
-import com.and04.naturealbum.data.dto.FirestoreUserWithStatus
-import com.and04.naturealbum.data.dto.FriendStatus
 
 @Composable
 fun MyPageSocialList(myFriends: List<FirebaseFriend>) {
@@ -76,156 +58,6 @@ fun MyPageSocialItem(myFriend: FirebaseFriend) {
 
     HorizontalDivider(thickness = 1.dp)
 }
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun MyPageSearch(
-    onSearchQueryChange: (String) -> Unit,
-    userWithStatusList: Map<String, FirestoreUserWithStatus>,
-    currentUid: String,
-    sendFriendRequest: (String, String) -> Unit,
-) {
-    var textFieldState by remember { mutableStateOf("") }
-    var expanded by rememberSaveable { mutableStateOf(false) }
-
-    Column(
-        Modifier.fillMaxSize()
-    ) {
-
-        // TODO: 키보드 올라왔을 때 검색 결과 부분까지 위로 올라가는 것 필요
-        RequestedList(
-            userWithStatusList = userWithStatusList,
-            currentUid = currentUid,
-            sendFriendRequest = sendFriendRequest
-        )
-
-        SearchBar(
-            modifier = Modifier
-                .align(Alignment.CenterHorizontally),
-            inputField = {
-                SearchBarDefaults.InputField(
-                    query = textFieldState,
-                    onSearch = { expanded = false },
-                    expanded = expanded,
-                    onExpandedChange = { expandedChange -> expanded = expandedChange },
-                    placeholder = {
-                        Text(text = stringResource(R.string.my_page_search_hint))
-                    },
-                    leadingIcon = {
-                        Icon(
-                            Icons.Default.Search,
-                            contentDescription = stringResource(R.string.my_page_search_bar_search_icon)
-                        )
-                    },
-                    onQueryChange = { query ->
-                        textFieldState = query
-                        onSearchQueryChange(query) // 검색 쿼리 변경 시 호출
-                    },
-                )
-            },
-            expanded = false,
-            onExpandedChange = {
-            },
-        ) {
-            // 검색 결과 리스트
-            RequestedList(
-                userWithStatusList = userWithStatusList,
-                currentUid = currentUid,
-                sendFriendRequest = sendFriendRequest
-            )
-        }
-
-    }
-}
-
-@Composable
-fun RequestedList(
-    userWithStatusList: Map<String, FirestoreUserWithStatus>,
-    currentUid: String,
-    sendFriendRequest: (String, String) -> Unit,
-) {
-    LazyColumn {
-        userWithStatusList.forEach { (uid, userWithStatus) ->
-            item(key = uid) {
-                RequestedItem(
-                    userWithStatus = userWithStatus,
-                    currentUid = currentUid,
-                    sendFriendRequest = sendFriendRequest
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun RequestedItem(
-    userWithStatus: FirestoreUserWithStatus,
-    currentUid: String,
-    sendFriendRequest: (String, String) -> Unit,
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(16.dp),
-        horizontalArrangement = Arrangement.spacedBy(16.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-
-        AsyncImage(
-            modifier = Modifier
-                .size(40.dp)
-                .clip(CircleShape),
-            contentScale = ContentScale.Crop,
-            model = userWithStatus.user.photoUrl,
-            contentDescription = stringResource(R.string.my_page_user_profile_image),
-        )
-
-        Column(
-            modifier = Modifier.weight(1f),
-            verticalArrangement = Arrangement.spacedBy(8.dp),
-            horizontalAlignment = Alignment.Start
-        ) {
-
-            Text(
-                text = userWithStatus.user.email,
-                style = MaterialTheme.typography.bodyLarge,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis
-            )
-
-            SuggestionChip(
-                onClick = {
-                    if (userWithStatus.status == FriendStatus.NORMAL) {
-                        sendFriendRequest(currentUid, userWithStatus.user.uid)
-                    }
-                },
-                enabled = userWithStatus.status == FriendStatus.NORMAL,
-                label = {
-                    val text = when (userWithStatus.status) {
-                        FriendStatus.SENT -> stringResource(R.string.my_page_friend_requested)
-                        FriendStatus.RECEIVED -> stringResource(R.string.my_page_friend_request_received)
-                        FriendStatus.FRIEND -> stringResource(R.string.my_page_friend)
-                        else -> stringResource(R.string.my_page_friend_request)
-                    }
-                    Text(text = text)
-                },
-                colors = SuggestionChipDefaults.suggestionChipColors(
-                    containerColor = when (userWithStatus.status) {
-                        FriendStatus.NORMAL -> MaterialTheme.colorScheme.primary
-                        else -> MaterialTheme.colorScheme.secondary
-                    },
-                    labelColor = when (userWithStatus.status) {
-                        FriendStatus.NORMAL -> MaterialTheme.colorScheme.onPrimary
-                        else -> MaterialTheme.colorScheme.onSecondary
-                    }
-                )
-            )
-        }
-    }
-
-    HorizontalDivider(thickness = 1.dp)
-}
-
 
 @Composable
 fun MyPageAlarm(
