+++ app/src/main/java/com/and04/naturealbum/ui/navigation/NatureAlbumState.kt
@@ -16,10 +16,12 @@ import androidx.compose.runtime.saveable.Saver
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.platform.LocalContext
 import androidx.core.content.FileProvider
+import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
 import com.and04.naturealbum.data.room.Label
 import com.and04.naturealbum.ui.LocationHandler
+import com.and04.naturealbum.ui.friend.FriendSearchScreen
 import com.and04.naturealbum.utils.ImageConvert
 import java.io.File
 
@@ -34,6 +36,7 @@ class NatureAlbumState(
     var fileName = mutableStateOf("")
     var selectedLabel = mutableStateOf<Label?>(null)
     private var imageFile = mutableStateOf<File?>(null)
+    var currentUid = mutableStateOf("")
 
     fun handleLauncher(result: ActivityResult) {
         if (result.resultCode == RESULT_OK) {
@@ -81,8 +84,7 @@ class NatureAlbumState(
         navController.popBackStack()
     }
 
-    fun getNavBackStackEntry() =
-        navController.getBackStackEntry(NavigateDestination.SavePhoto.route)
+    fun getNavBackStackEntry(route: String) = navController.getBackStackEntry(route)
 
     fun navigateToAlbum() = navController.navigate(NavigateDestination.Album.route)
 
@@ -108,6 +110,8 @@ class NatureAlbumState(
     fun navigateToAlbumInfo(photoDetailId: Int) {
         navController.navigate("${NavigateDestination.PhotoInfo.route}/$photoDetailId")
     }
+
+    fun navigateToFriendSearch() = navController.navigate(NavigateDestination.FriendSearch.route)
 }
 
 @Composable
